backend/tests/requirements.txt
## Testing
absl-py==1.0.0
flake8==3.5.0
Flask-Testing==0.7.1
freezegun==0.3.10
mock==2.0.0
pytest==7.1.2
pytest-cov==3.0.0

setuptools~=57.4.0